refactor: set VIP number for VIP-403
