+++ simulations/vip-403/arbitrumsepolia.ts
@@ -4,14 +4,18 @@ import { NETWORK_ADDRESSES } from "src/networkAddresses";
 import { expectEvents } from "src/utils";
 import { forking, testForkedNetworkVipCommands } from "src/vip-framework";
 
-import vip405, { ARBITRUM_ACM, ARBITRUM_ACM_AGGREGATOR, DEFAULT_ADMIN_ROLE } from "../../vips/vip-405/bscmainnet";
+import vip403, {
+  ARBITRUMSEPOLIA_ACM_AGGREGATOR,
+  ARBITRUM_SEPOLIA_ACM,
+  DEFAULT_ADMIN_ROLE,
+} from "../../vips/vip-403/bsctestnet";
 import ACM_COMMANDS_AGGREGATOR_ABI from "./abi/ACMCommandsAggregator.json";
 import ACCESS_CONTROL_MANAGER_ABI from "./abi/AccessControlManager.json";
 
-const { arbitrumone } = NETWORK_ADDRESSES;
+const { arbitrumsepolia } = NETWORK_ADDRESSES;
 
-forking(280572842, async () => {
-  testForkedNetworkVipCommands("VIP 405 Multichain Governance - Permissions", await vip405(), {
+forking(101407909, async () => {
+  testForkedNetworkVipCommands("VIP 403 Multichain Governance - Permissions", await vip403(), {
     callbackAfterExecution: async txResponse => {
       await expectEvents(txResponse, [ACCESS_CONTROL_MANAGER_ABI], ["PermissionGranted"], [2]);
       await expectEvents(txResponse, [ACCESS_CONTROL_MANAGER_ABI], ["PermissionRevoked"], [4]);
@@ -21,19 +25,19 @@ forking(280572842, async () => {
   });
 
   describe("Post-VIP behaviour", async () => {
-    const acm = new ethers.Contract(ARBITRUM_ACM, ACCESS_CONTROL_MANAGER_ABI, ethers.provider);
+    const acm = new ethers.Contract(ARBITRUM_SEPOLIA_ACM, ACCESS_CONTROL_MANAGER_ABI, ethers.provider);
     it("check if DEFAULT_ROLE has been revoked for ACMAggregator", async () => {
-      expect(await acm.hasRole(DEFAULT_ADMIN_ROLE, ARBITRUM_ACM_AGGREGATOR)).to.be.false;
+      expect(await acm.hasRole(DEFAULT_ADMIN_ROLE, ARBITRUMSEPOLIA_ACM_AGGREGATOR)).to.be.false;
     });
 
     it("check few permissions", async () => {
       const role1 = ethers.utils.solidityPack(
         ["address", "string"],
-        [arbitrumone.REDSTONE_ORACLE, "setTokenConfig(TokenConfig)"],
+        [arbitrumsepolia.REDSTONE_ORACLE, "setTokenConfig(TokenConfig)"],
       );
 
       const roleHash = ethers.utils.keccak256(role1);
-      expect(await acm.hasRole(roleHash, arbitrumone.GUARDIAN)).to.be.true;
+      expect(await acm.hasRole(roleHash, arbitrumsepolia.GUARDIAN)).to.be.true;
     });
   });
 });
