+++ simulations/vip-403/bsctestnet.ts
@@ -1,12 +1,12 @@
 import { expectEvents } from "../../src/utils";
 import { forking, testVip } from "../../src/vip-framework";
-import vip405 from "../../vips/vip-405/bsctestnet";
+import vip403 from "../../vips/vip-403/bsctestnet";
 import ACM_COMMANDS_AGGREGATOR_ABI from "./abi/ACMCommandsAggregator.json";
 import ACCESS_CONTROL_MANAGER_ABI from "./abi/AccessControlManager.json";
 import OMNICHAIN_PROPOSAL_SENDER_ABI from "./abi/OmnichainProposalSender.json";
 
 forking(46024012, async () => {
-  testVip("VIP 405 Multichain Governance - Permissions", await vip405(), {
+  testVip("VIP 403 Multichain Governance - Permissions", await vip403(), {
     callbackAfterExecution: async txResponse => {
       await expectEvents(
         txResponse,
