+++ simulations/vip-403/ethereum.ts
@@ -4,14 +4,14 @@ import { NETWORK_ADDRESSES } from "src/networkAddresses";
 import { expectEvents } from "src/utils";
 import { forking, testForkedNetworkVipCommands } from "src/vip-framework";
 
-import vip405, { DEFAULT_ADMIN_ROLE, ETHEREUM_ACM, ETHEREUM_ACM_AGGREGATOR } from "../../vips/vip-405/bscmainnet";
+import vip403, { DEFAULT_ADMIN_ROLE, ETHEREUM_ACM, ETHEREUM_ACM_AGGREGATOR } from "../../vips/vip-403/bscmainnet";
 import ACM_COMMANDS_AGGREGATOR_ABI from "./abi/ACMCommandsAggregator.json";
 import ACCESS_CONTROL_MANAGER_ABI from "./abi/AccessControlManager.json";
 
 const { ethereum } = NETWORK_ADDRESSES;
 
 forking(21327971, async () => {
-  testForkedNetworkVipCommands("VIP 405 Multichain Governance - Permissions", await vip405(), {
+  testForkedNetworkVipCommands("VIP 403 Multichain Governance - Permissions", await vip403(), {
     callbackAfterExecution: async txResponse => {
       await expectEvents(txResponse, [ACCESS_CONTROL_MANAGER_ABI], ["PermissionGranted"], [2]);
       await expectEvents(txResponse, [ACCESS_CONTROL_MANAGER_ABI], ["PermissionRevoked"], [4]);
