+++ simulations/vip-403/opbnbmainnet.ts
@@ -4,36 +4,34 @@ import { NETWORK_ADDRESSES } from "src/networkAddresses";
 import { expectEvents } from "src/utils";
 import { forking, testForkedNetworkVipCommands } from "src/vip-framework";
 
-import vip405, { DEFAULT_ADMIN_ROLE, SEPOLIA_ACM, SEPOLIA_ACM_AGGREGATOR } from "../../vips/vip-405/bsctestnet";
+import vip403, { DEFAULT_ADMIN_ROLE, OPBNB_ACM, OPBNB_ACM_AGGREGATOR } from "../../vips/vip-403/bscmainnet";
 import ACM_COMMANDS_AGGREGATOR_ABI from "./abi/ACMCommandsAggregator.json";
 import ACCESS_CONTROL_MANAGER_ABI from "./abi/AccessControlManager.json";
 
-const { sepolia } = NETWORK_ADDRESSES;
+const { opbnbmainnet } = NETWORK_ADDRESSES;
 
-forking(7169767, async () => {
-  testForkedNetworkVipCommands("VIP 405 Multichain Governance - Permissions", await vip405(), {
+forking(41392857, async () => {
+  testForkedNetworkVipCommands("VIP 403 Multichain Governance - Permissions", await vip403(), {
     callbackAfterExecution: async txResponse => {
-      await expectEvents(txResponse, [ACCESS_CONTROL_MANAGER_ABI], ["PermissionGranted"], [2]);
       await expectEvents(txResponse, [ACCESS_CONTROL_MANAGER_ABI], ["PermissionRevoked"], [4]);
-      await expectEvents(txResponse, [ACM_COMMANDS_AGGREGATOR_ABI], ["GrantPermissionsExecuted"], [1]);
       await expectEvents(txResponse, [ACM_COMMANDS_AGGREGATOR_ABI], ["RevokePermissionsExecuted"], [1]);
     },
   });
 
   describe("Post-VIP behaviour", async () => {
-    const acm = new ethers.Contract(SEPOLIA_ACM, ACCESS_CONTROL_MANAGER_ABI, ethers.provider);
+    const acm = new ethers.Contract(OPBNB_ACM, ACCESS_CONTROL_MANAGER_ABI, ethers.provider);
     it("check if DEFAULT_ROLE has been revoked for ACMAggregator", async () => {
-      expect(await acm.hasRole(DEFAULT_ADMIN_ROLE, SEPOLIA_ACM_AGGREGATOR)).to.be.false;
+      expect(await acm.hasRole(DEFAULT_ADMIN_ROLE, OPBNB_ACM_AGGREGATOR)).to.be.false;
     });
 
     it("check few permissions", async () => {
       const role1 = ethers.utils.solidityPack(
         ["address", "string"],
-        [sepolia.REDSTONE_ORACLE, "setTokenConfig(TokenConfig)"],
+        [opbnbmainnet.XVS_VAULT_PROXY, "add(address,uint256,address,uint256,uint256)"],
       );
 
       const roleHash = ethers.utils.keccak256(role1);
-      expect(await acm.hasRole(roleHash, sepolia.GUARDIAN)).to.be.true;
+      expect(await acm.hasRole(roleHash, opbnbmainnet.GUARDIAN)).to.be.false;
     });
   });
 });
