+++ simulations/vip-403/opbnbtestnet.ts
@@ -4,14 +4,18 @@ import { NETWORK_ADDRESSES } from "src/networkAddresses";
 import { expectEvents } from "src/utils";
 import { forking, testForkedNetworkVipCommands } from "src/vip-framework";
 
-import vip405, { DEFAULT_ADMIN_ROLE, OPBNB_ACM, OPBNB_ACM_AGGREGATOR } from "../../vips/vip-405/bscmainnet";
+import vip403, {
+  DEFAULT_ADMIN_ROLE,
+  OPBNBTESTNET_ACM,
+  OPBNBTESTNET_ACM_AGGREGATOR,
+} from "../../vips/vip-403/bsctestnet";
 import ACM_COMMANDS_AGGREGATOR_ABI from "./abi/ACMCommandsAggregator.json";
 import ACCESS_CONTROL_MANAGER_ABI from "./abi/AccessControlManager.json";
 
-const { opbnbmainnet } = NETWORK_ADDRESSES;
+const { opbnbtestnet } = NETWORK_ADDRESSES;
 
-forking(41392857, async () => {
-  testForkedNetworkVipCommands("VIP 405 Multichain Governance - Permissions", await vip405(), {
+forking(45916044, async () => {
+  testForkedNetworkVipCommands("VIP 403 Multichain Governance - Permissions", await vip403(), {
     callbackAfterExecution: async txResponse => {
       await expectEvents(txResponse, [ACCESS_CONTROL_MANAGER_ABI], ["PermissionRevoked"], [4]);
       await expectEvents(txResponse, [ACM_COMMANDS_AGGREGATOR_ABI], ["RevokePermissionsExecuted"], [1]);
@@ -19,19 +23,19 @@ forking(41392857, async () => {
   });
 
   describe("Post-VIP behaviour", async () => {
-    const acm = new ethers.Contract(OPBNB_ACM, ACCESS_CONTROL_MANAGER_ABI, ethers.provider);
+    const acm = new ethers.Contract(OPBNBTESTNET_ACM, ACCESS_CONTROL_MANAGER_ABI, ethers.provider);
     it("check if DEFAULT_ROLE has been revoked for ACMAggregator", async () => {
-      expect(await acm.hasRole(DEFAULT_ADMIN_ROLE, OPBNB_ACM_AGGREGATOR)).to.be.false;
+      expect(await acm.hasRole(DEFAULT_ADMIN_ROLE, OPBNBTESTNET_ACM_AGGREGATOR)).to.be.false;
     });
 
     it("check few permissions", async () => {
       const role1 = ethers.utils.solidityPack(
         ["address", "string"],
-        [opbnbmainnet.XVS_VAULT_PROXY, "add(address,uint256,address,uint256,uint256)"],
+        [opbnbtestnet.XVS_VAULT_PROXY, "add(address,uint256,address,uint256,uint256)"],
       );
 
       const roleHash = ethers.utils.keccak256(role1);
-      expect(await acm.hasRole(roleHash, opbnbmainnet.GUARDIAN)).to.be.false;
+      expect(await acm.hasRole(roleHash, opbnbtestnet.GUARDIAN)).to.be.false;
     });
   });
 });
